Create LICENSE
Copyright 2025 James Phillip Palomares, Palomares Group

Licensed under the Apache License, Version 2.0 (the "License");
you may not use this work except in compliance with the License.
You may obtain a copy of the License at:

    http://www.apache.org/licenses/LICENSE-2.0

Unless required by applicable law or agreed to in writing, software
distributed under the License is distributed on an "AS IS" BASIS,
WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
See the License for the specific language governing permissions and
limitations under the License.

This license applies to all original content, code, documentation, and theory published in this repository, including but not limited to CNFT and CAG frameworks. Attribution to James Phillip Palomares and Palomares Group is required in all derivative works and redistributions.

Patent rights are granted under this license. Any use of the CNFT or CAG mechanisms must include proper attribution and comply with the terms herein.


Signed-off-by: charliebrownheh25-dotcom <charliebrownheh25@gmail.com>